+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Fixed
+
+- add the /p parameter to the productList structured data URL
+
 ## [0.7.0] - 2021-08-05
 ### Added 
 - `settingsSchema` to allow stores to send `pricesWithTax` and specify the number of `decimals`